django-rest/requirements.txt
asgiref==3.5.2
Babel==2.10.3
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
django-filter==22.1
django-money==3.0.0
djangorestframework==3.13.1
Pillow==9.2.0
py-moneyed==2.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
